# Conflicts:
#	AnotherOneBitesTheCrust/Game.cpp
#	AnotherOneBitesTheCrust/Renderable.cpp
#	AnotherOneBitesTheCrust/RenderingEngine.cpp

+++ AnotherOneBitesTheCrust/ContentLoading.cpp
@@ -13,7 +13,7 @@ bool loadVehicleData(char* filename, Vehicle* vehicle) {
 	rapidjson::Document d;
 	d.ParseStream(reader);
 	if (d.HasMember("mass")) {
-		vehicle->chassisMass = d["mass"].GetDouble();
+		vehicle->chassisMass = (float)d["mass"].GetDouble();
 	}
 	vehicle->updateTuning();
